+++ src/components/CardWrapper.tsx
@@ -12,8 +12,9 @@ import {
 } from '../lib/thresholdCalulationFunc';
 import { FlipCard } from './FlipCard';
 import { FormattedStringVariables } from '../interfaces/formattedStringVariables';
-import { provideFormattedStringVariables } from '../lib/formattedString';
+import { compileFormattedString, provideFormattedStringVariables } from '../lib/formattedString';
 import { mappingMetricUnitName } from '../lib/metricUnitMapping';
+import { FlipCardNoData } from './FlipCardNoData';
 
 interface CardWrapperProps {
   index: number;
@@ -38,10 +39,6 @@ export const CardWrapper: React.FC<CardWrapperProps> = ({
   flipped,
   index,
 }) => {
-  const cardTitle = getTitle(options, fieldsConfig, series);
-  const cardSubtitle = getSubtitle(options, fieldsConfig, series);
-  const cardUrl = getUrl(options, fieldsConfig, series);
-  const cardUrlTargetBlank = getUrlTargetBlank(options, fieldsConfig, series);
   // Calculate values depending on fieldsConfig and override fields
   const metricUnit = getMetricUnit(
     fieldsConfig.defaults.custom.displayValueMetric,
@@ -49,7 +46,7 @@ export const CardWrapper: React.FC<CardWrapperProps> = ({
     series,
     fieldsConfig.overrides
   );
-
+  // Get formatted string variables (like query name, query value, etc.)
   const stringFormattedVariables: FormattedStringVariables = provideFormattedStringVariables(
     index,
     series,
@@ -57,7 +54,37 @@ export const CardWrapper: React.FC<CardWrapperProps> = ({
     queryValue,
     metricUnit || ''
   );
+
+  // Debug, help people to show what their can do with the variables
+  console.log('stringFormattedVariables', stringFormattedVariables);
+
+  const cardTitle = compileFormattedString(getTitle(options, fieldsConfig, series), stringFormattedVariables);
+  const cardSubtitle = compileFormattedString(getSubtitle(options, fieldsConfig, series), stringFormattedVariables);
+  const cardUrl = compileFormattedString(getUrl(options, fieldsConfig, series), stringFormattedVariables);
+  const cardUrlTargetBlank = getUrlTargetBlank(options, fieldsConfig, series);
+
   const thresholdsConf = getThresholdsConf(fieldsConfig, series);
+
+  if (queryValue === undefined || queryValue === null) {
+    return (
+      <>
+        {!options.isNothingOnNoData && (
+          <FlipCardNoData
+            width={cardWidth}
+            height={cardHeight}
+            title={cardTitle}
+            subtitle={cardSubtitle}
+            url={cardUrl}
+            urlTargetBlank={cardUrlTargetBlank}
+            showMetric={fieldsConfig.defaults.custom.displayValueMetric}
+            fontStyle={fieldsConfig.defaults.custom.fontFormat}
+            cornerRadius={options.cornerRadius}
+            isFlipped={flipped}
+          />
+        )}
+      </>
+    );
+  }
 
   return (
     <FlipCard
@@ -71,7 +98,6 @@ export const CardWrapper: React.FC<CardWrapperProps> = ({
       metricUnit={metricUnit}
       fontStyle={fieldsConfig.defaults.custom.fontFormat}
       thresholds={thresholdsConf}
-      formattedVariables={stringFormattedVariables}
       value={queryValue}
       isFlipped={flipped}
       cornerRadius={options.cornerRadius}
@@ -83,7 +109,7 @@ interface CardWrapperPropsAggregateQuery {
   data: PanelData;
   fieldsConfig: FieldConfigSource<any>;
   options: StatusPanelOptions;
-  queriesValuesAggregated: number[];
+  queriesValuesAggregated: number[][];
   cardWidth: number;
   cardHeight: number;
   flipped: boolean;
@@ -102,29 +128,47 @@ export const CardWrapperAggregateQuery: React.FC<CardWrapperPropsAggregateQuery>
   const thresholdsConf = fieldsConfig.defaults.custom.thresholds;
   // For all aggregateQueriesValues, get actual threshold
   let actualThresholds = [];
-  for (let queryValue of queriesValuesAggregated) {
-    actualThresholds.push(getActualThreshold(thresholdsConf, queryValue));
+  for (let queryValues of queriesValuesAggregated) {
+    // note: queryValues is already the aggregated value for each dataframe of a query
+    for (let queryValue of queryValues) {
+      actualThresholds.push(getActualThreshold(thresholdsConf, queryValue));
+    }
   }
+
   // Get index of the worst threshold
   const actualThreshold = actualThresholds.reduce((prev, current) =>
     (prev.value || 0) > (current.value || 0) ? prev : current
   );
   const thresholdIndex = actualThresholds.indexOf(actualThreshold);
 
-  const queryValue = queriesValuesAggregated[thresholdIndex];
+  const queryValue = queriesValuesAggregated.flat()[thresholdIndex];
   const metricUnit = mappingMetricUnitName(fieldsConfig.defaults.custom.metricUnit);
-  const cardTitle = getTitle(options, fieldsConfig, data.series[thresholdIndex]);
-  const cardSubtitle = getSubtitle(options, fieldsConfig, data.series[thresholdIndex]);
-  const cardUrl = getUrl(options, fieldsConfig, data.series[thresholdIndex]);
-  const cardUrlTargetBlank = getUrlTargetBlank(options, fieldsConfig, data.series[thresholdIndex]);
 
   const stringFormattedVariables: FormattedStringVariables = provideFormattedStringVariables(
     thresholdIndex,
     data.series[thresholdIndex],
     data,
     queryValue,
-    metricUnit
+    metricUnit,
+    true
   );
+  // Debug, help people to show what their can do with the variables
+  console.log('stringFormattedVariables', stringFormattedVariables);
+
+  // Formatted Texts
+  const cardTitle = compileFormattedString(
+    getTitle(options, fieldsConfig, data.series[thresholdIndex]),
+    stringFormattedVariables
+  );
+  const cardSubtitle = compileFormattedString(
+    getSubtitle(options, fieldsConfig, data.series[thresholdIndex]),
+    stringFormattedVariables
+  );
+  const cardUrl = compileFormattedString(
+    getUrl(options, fieldsConfig, data.series[thresholdIndex]),
+    stringFormattedVariables
+  );
+  const cardUrlTargetBlank = getUrlTargetBlank(options, fieldsConfig, data.series[thresholdIndex]);
 
   return (
     <FlipCard
@@ -139,7 +183,6 @@ export const CardWrapperAggregateQuery: React.FC<CardWrapperPropsAggregateQuery>
       fontStyle={fieldsConfig.defaults.custom.fontFormat}
       cornerRadius={options.cornerRadius}
       thresholds={fieldsConfig.defaults.custom.thresholds}
-      formattedVariables={stringFormattedVariables}
       value={queryValue}
       isFlipped={flipped}
     />